Processing/requirements.txt
flask==3.0.3
connexion==3.1.0
APScheduler==3.10.4
MarkupSafe>=2.1.1
requests>=2.27
PyYAML>=5.1
connexion[flask]==3.1.0
connexion[uvicorn]==3.1.0
connexion[swagger-ui]==3.1.0
starlette>=0.14.2
